refactor: DeliveryMethod 패키지 변경
