+++ src/main/kotlin/com/petqua/domain/delivery/DeliveryMethod.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.cart
+package com.petqua.domain.delivery
 
 import com.petqua.exception.cart.CartProductException
 import com.petqua.exception.cart.CartProductExceptionType.INVALID_DELIVERY_METHOD
@@ -12,7 +12,7 @@ enum class DeliveryMethod(
     SAFETY("안전 운송"),
     PICK_UP("직접 방문"),
     ;
-    
+
     companion object {
         fun from(name: String): DeliveryMethod {
             return enumValues<DeliveryMethod>().find { it.name == name.uppercase(ENGLISH) }
